requirements.txt
ipython
Pillow
protobuf
python-dotenv
streamlit
google-generativeai 